tests/wpt/tests/navigation-api/navigate-event/navigate-form.html

<!doctype html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<form id="form" method="post" action=""></form>
<script>
async_test(t => {
  navigation.onnavigate = t.step_func_done(e => {
    e.preventDefault();

    assert_equals(e.navigationType, "replace");
    assert_true(e.cancelable);
    assert_true(e.canIntercept);
    assert_false(e.userInitiated);
    assert_false(e.hashChange);
    assert_equals(e.downloadRequest, null);
    assert_equals(e.destination.url, location.href);
    assert_false(e.destination.sameDocument);
    assert_equals(e.destination.key, "");
    assert_equals(e.destination.id, "");
    assert_equals(e.destination.index, -1);
    assert_not_equals(e.formData, null);
    // NavigateEvent sourceElement is still in development, so test whether it is available.
    if ("sourceElement" in e) assert_equals(e.sourceElement, form);
  });
  window.onload = t.step_func(() => form.submit());
}, "<form> submission fires navigate event");
</script>
