LayoutTests/imported/w3c/web-platform-tests/navigation-api/navigate-event/navigate-form-expected.txt


FAIL <form> submission fires navigate event assert_equals: expected "replace" but got "push"
